10/02:micro:00

+++ src/draw/project.c
@@ -4,30 +4,30 @@ t_cam	ft_project(t_model model, float rad)
 {
 	t_cam	cam;
 
-	cam.x = ((float)(model.x - model.y) * 50) * cos(rad) + 500;
-	cam.y = ((float)((model.x + model.y) * 50) * sin(rad) - (model.z * 20)) + 200;
+	cam.x = ((model.x - model.y) * 50) * cos(rad) + 500;
+	cam.y = (((model.x + model.y) * 50) * sin(rad) - (model.z * 20)) + 200;
 	return (cam);
 }
 
 void	ft_rotx(t_model *model, float rad)
 {
 	model->x = round((model->x * 1) + (model->x * 0) + (model->x * 0));
-	model->z = round(((float)model->z * 0) + ((float)model->z * cos(rad)) + ((float)model->z * -sin(rad)));
-	model->y = round((float)(model->y * 0) + (float)(model->y * sin(rad)) + (float)(model->y * cos(rad)));
+	model->z = round((model->z * 0) + (model->z * cos(rad)) + (model->z * -sin(rad)));
+	model->y = round((model->y * 0) + (model->y * sin(rad)) + (model->y * cos(rad)));
 }
 
 void	ft_rotz(t_model *model, float rad)
 {
-	model->x = round(((float)model->x * cos(rad)) + (float)(model->x * 0) + ((float)model->x * sin(rad)));
+	model->x = round((model->x * cos(rad)) + (model->x * 0) + (model->x * sin(rad)));
 	model->z = (model->z * 0) + (model->z * 1) + (model->z * 0);
-	model->y = round(((float)model->y * -sin(rad)) + (float)(model->y * 0) + ((float)model->y * cos(rad)));
+	model->y = round((model->y * -sin(rad)) + (model->y * 0) + (model->y * cos(rad)));
 }
 
 void	ft_roty(t_model *model, float rad)
 {
-	model->x = round(((float)model->x * cos(rad)) + ((float)model->x * -sin(rad)) + (float)(model->x * 0));
-	model->z = round(((float)model->z * sin(rad)) + ((float)model->z * cos(rad)) + (float)(model->z * 0));
-	model->y = round(((float)model->y * 0) + ((float)model->y * 0) + (float)(model->y * 1));
+	model->x = round((model->x * cos(rad)) + (model->x * -sin(rad)) + (model->x * 0));
+	model->z = round((model->z * sin(rad)) + (model->z * cos(rad)) + (model->z * 0));
+	model->y = round((model->y * 0) + (model->y * 0) + (model->y * 1));
 }
 
 // x = 	1 0 		0
